Report code cleanup - Separate Test Report Steps
