+++ psm-app/integration-tests/src/test/java/gov/medicaid/features/report/steps/ProviderTypesReportStepDefinitions.java
@@ -1,0 +1,50 @@
+package gov.medicaid.features.report.steps;
+
+import cucumber.api.java.en.Given;
+import cucumber.api.java.en.Then;
+
+import net.thucydides.core.annotations.Steps;
+
+public class ProviderTypesReportStepDefinitions {
+    @Steps
+    private ProviderTypesReportSteps providerTypesSteps;
+
+    @Then("^I should see the provider types page$")
+    public void i_should_see_the_provider_types_page() {
+        providerTypesSteps.checkOnProviderTypesPage();
+    }
+
+    @Then("^I should see a provider types table with '(\\d+)' enrollments for '(.*)' in month '(.*)'$")
+    public void i_should_see_a_provider_types_table_with_enrollments_for_type_in_month(
+        int numEnrollments,
+        String providerTypeDesc,
+        String month
+    ) {
+        providerTypesSteps.checkProviderTypesPageHasEnrollmentsForMonth(numEnrollments, providerTypeDesc, month);
+    }
+
+    @Then("^I should see no results for month '(.*)'$")
+    public void i_should_see_no_results_for_month(String month) {
+        providerTypesSteps.checkProviderTypesPageHasNoResultsForMonth(month);
+    }
+
+    @Given("^I select the provider type '(.*)'$")
+    public void i_select_the_provider_type(String providerTypeDesc) {
+        providerTypesSteps.selectProviderTypeOption(providerTypeDesc);
+    }
+
+    @Given("^I filter for provider types$")
+    public void i_filter_for_provider_types() {
+        providerTypesSteps.filterProviderTypes();
+    }
+
+    @Given("^I download the provider types report$")
+    public void i_download_the_provider_types_report() {
+        providerTypesSteps.downloadProviderTypesReportCsv();
+    }
+
+    @Then("^I should see provider type '(.*)' selected$")
+    public void i_should_see_provider_type_selected(String providerTypeDesc) {
+        providerTypesSteps.checkProviderTypeSelected(providerTypeDesc);
+    }
+}
